backend/requirements.txt
# OpenContextGraph Backend Dependencies

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Authentication (OIDC)
python-jose[cryptography]>=3.3.0
httpx>=0.26.0

# Azure Services (for VoiceLive and Blob Storage)
azure-identity>=1.15.0
azure-core>=1.29.5
azure-storage-blob>=12.19.0

# Azure AI VoiceLive (Real-time Voice) - Optional
# azure-ai-voicelive[aiohttp]>=1.0.0  # Uncomment if VoiceLive is needed

# WebRTC for Avatar Video (Optional)
# aiortc>=1.6.0  # Uncomment if Avatar video is needed

# Agent Framework
langchain>=0.1.0
langgraph>=0.0.20

# LLM Providers
openai>=1.0.0
# Azure AI Foundry (Agent Service) - Pinning for Project Endpoint Support
azure-ai-projects==1.0.0b13
azure-ai-agents>=1.0.0b1
# google-genai>=0.1.0  # Optional: Gemini

# ETL & Document Processing
unstructured[all-docs]>=0.12.0
python-magic>=0.4.27

# Memory & Orchestration
zep-python>=2.0.0
temporalio>=1.4.0

# Knowledge Graph
graphiti-core>=0.2.0
networkx>=3.0.0

# Database
asyncpg>=0.29.0
sqlalchemy>=2.0.0

# Observability
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-exporter-otlp-proto-grpc>=1.22.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
Pillow>=10.2.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
ruff>=0.1.0
